PassCarr Choosing seat functionality added

+++ Lab06.MVC.Carriage.BL/Model/OrderModel.cs
@@ -8,6 +8,6 @@
 
         public string UserId { get; set; }
 
-        public byte SeatNumber { get; set; }
+        public int SeatNumber { get; set; }
     }
 }
